+++ grafana-plugin/src/models/alert_receive_channel_filters/alert_receive_channel_filters.ts
@@ -20,6 +20,7 @@ export class AlertReceiveChannelFiltersStore extends BaseStore {
     this.path = '/alert_receive_channels/';
   }
 
+  @action.bound
   getSearchResult() {
     if (!this.searchResult) {
       return undefined;
@@ -28,7 +29,7 @@ export class AlertReceiveChannelFiltersStore extends BaseStore {
     return this.searchResult.map((value: SelectOption['value']) => this.items?.[value]);
   }
 
-  @action
+  @action.bound
   async updateItems(query = '') {
     const results = await makeRequest(`${this.path}`, {
       params: { search: query, filters: true },