+++ grafana-plugin/src/models/loader/loader.ts
@@ -12,7 +12,7 @@ class LoaderStoreClass {
     makeObservable(this);
   }
 
-  @action
+  @action.bound
   setLoadingAction(actionKey: string, isLoading: boolean) {
     this.items[actionKey] = isLoading;
   }